backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
h5py==3.10.0
numpy==1.25.2
Pillow==10.1.0
pandas==2.1.3

# Caching and database
redis==5.0.1
aioredis==2.0.1

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing and utilities
opencv-python-headless==4.8.1.78
scikit-image==0.22.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# No additional CORS package needed - FastAPI has built-in CORS support
